+++ src/enginefactory.hpp
@@ -16,7 +16,7 @@ public:
   enginefactory& with_gravity(float gravity) noexcept;
   enginefactory& with_fullscreen(bool fullscreen) noexcept;
   enginefactory& with_sentry(std::string_view dsn) noexcept;
-  enginefactory& with_ticks(int ticks) noexcept;
+  enginefactory& with_ticks(uint8_t ticks) noexcept;
 
   std::shared_ptr<engine> create() const;
 
@@ -27,5 +27,5 @@ private:
   float _scale{1.0};
   float _gravity{9.8f};
   bool _fullscreen{false};
-  int _ticks{0};
+  uint8_t _ticks{0};
 };
